+++ CHANGES.md
@@ -1,6 +1,18 @@
 twilio-java changelog
 =====================
 
+[2023-06-01] Version 9.6.2
+--------------------------
+**Api**
+- Add `Trim` to create Conference Participant API
+
+**Intelligence**
+- First public beta release for Voice Intelligence APIs with client libraries
+
+**Messaging**
+- Add new `errors` attribute to us_app_to_person resource. This attribute will provide additional information about campaign registration errors.
+
+
 [2023-05-18] Version 9.6.1
 --------------------------
 **Library - Fix**